// src/components/Widgets/widget.types.ts
export type HTMLTag =
    | 'div'
    | 'span'
    | 'header'
    | 'ul'
    | 'ol'
    | 'li'
    | 'nav'
    | 'a'
    | 'img'
    | 'h1'
    | 'h2'
    | 'h3'
    | 'h4'
    | 'h5'
    | 'h6'
    | 'p'
    | 'button'
    | 'i'
    | 'section'
    | 'main'
    | 'aside'

export type BorderColor = 'red' | 'green' | 'blue' | 'yellow' | 'white' | 'black' | 'purple'